fix: remove unused consts and fns
