+++ src/components/ControlBar/Chips.js
@@ -2,13 +2,13 @@ import React from 'react'
 import { connect } from 'react-redux'
 import PropTypes from 'prop-types'
 import cx from 'classnames'
-import { useWindowDimensions } from '../WindowDimensionsProvider'
-import { isSmallScreen } from '../../modules/smallScreen'
 import Chip from './DashboardItemChip'
 import { sGetAllDashboards } from '../../reducers/dashboards'
 import { sGetDashboardsFilter } from '../../reducers/dashboardsFilter'
 import { sGetSelectedId } from '../../reducers/selected'
-import { getFilteredDashboards } from './DashboardsBar'
+import { useWindowDimensions } from '../WindowDimensionsProvider'
+import { isSmallScreen } from '../../modules/smallScreen'
+import { getFilteredDashboards } from '../../modules/getFilteredDashboards'
 
 import classes from './styles/Chips.module.css'
