+++ src/components/ControlBar/styles/EditBar.module.css
@@ -7,6 +7,7 @@
     box-shadow: rgba(0, 0, 0, 0.2) 0px 0px 6px 3px;
     position: relative;
     z-index: 10;
+    flex: none;
 }
 
 .controls {
