+++ src/components/Dashboard/styles/EditDashboard.module.css
@@ -1,0 +1,6 @@
+.container {
+    display: flex;
+    flex-direction: column;
+    height: 100%;
+    position: relative;
+}
